+++ packages/app/ui/components/AppSetting.tsx
@@ -3,7 +3,7 @@ import { Icon } from '@tloncorp/ui';
 import { Pressable } from '@tloncorp/ui';
 import { useCallback, useState } from 'react';
 
-import { ListItem } from './ListItem';
+import { ListItem } from './listItems';
 
 interface Props {
   title: string;
