+++ packages/app/ui/components/AuthorRow.tsx
@@ -9,8 +9,8 @@ import { ContactAvatar } from './Avatar';
 import { Badge } from './Badge';
 import { ChatMessageDeliveryStatus } from './ChatMessage/ChatMessageDeliveryStatus';
 import { ContactName } from './ContactNameV2';
-import { useBoundHandler } from './ListItem/listItemUtils';
 import { SentTimeText } from './SentTimeText';
+import { useBoundHandler } from './listItems/listItemUtils';
 
 const RoleBadge = View.styleable<{ role: string }>(
   ({ role, ...props }, ref) => {
