+++ packages/app/ui/components/ChatOptionsSheet.tsx
@@ -25,8 +25,8 @@ import {
   ActionSheet,
   createActionGroups,
 } from './ActionSheet';
-import { ListItem } from './ListItem';
 import { useNotificationLevelOptions } from './NotificationLevelSelector';
+import { ListItem } from './listItems';
 
 function getNotificationTitle(
   volumeSettings: { level: ub.NotificationLevel } | null | undefined,
