+++ packages/app/ui/components/ContactRow.tsx
@@ -7,7 +7,7 @@ import { Stack, View, XStack } from 'tamagui';
 
 import { triggerHaptic } from '../utils';
 import { ContactName } from './ContactNameV2';
-import { ListItem } from './ListItem';
+import { ListItem } from './listItems';
 
 function ContactRowItemRaw({
   contact,
