+++ packages/app/ui/components/GroupSelector.tsx
@@ -10,7 +10,7 @@ import {
 import { Stack, View } from 'tamagui';
 
 import { AlphaSegmentedGroups } from '../hooks/groupsSorters';
-import { GroupListItem, ListItem } from './ListItem';
+import { GroupListItem, ListItem } from './listItems';
 
 interface Props {
   onSelect?: (group: db.Group) => void;
