+++ packages/app/ui/components/Tabs.tsx
@@ -2,7 +2,7 @@ import { Text } from '@tloncorp/ui';
 import { ReactNode } from 'react';
 import { XStack, styled, withStaticProperties } from 'tamagui';
 
-import { useBoundHandler } from './ListItem/listItemUtils';
+import { useBoundHandler } from './listItems/listItemUtils';
 
 const TabsWrapper = styled(XStack, {
   width: '100%',
