+++ packages/app/ui/components/UserProfileScreenView.tsx
@@ -30,17 +30,17 @@ import {
 } from 'tamagui';
 
 import { RootStackParamList } from '../../navigation/types';
-import { useNavigation as useContextNavigation } from '../contexts/navigation';
 import { useContact, useCurrentUserId } from '../contexts/appDataContext';
+import { useNavigation as useContextNavigation } from '../contexts/navigation';
 import { useGroupTitle } from '../utils';
 import { ContactAvatar, GroupAvatar } from './Avatar';
 import { ContactName } from './ContactNameV2';
-import { useBoundHandler } from './ListItem/listItemUtils';
 import {
   PhoneAttestDisplay,
   TwitterAttestDisplay,
 } from './Profile/ConnectedAccountsWidget';
 import { WidgetPane } from './WidgetPane';
+import { useBoundHandler } from './listItems/listItemUtils';
 
 interface Props {
   userId: string;
