+++ packages/app/ui/index.tsx
@@ -40,7 +40,7 @@ export * from './components/GroupMembersScreenView';
 export * from './components/GroupPreviewSheet';
 export * from './components/InviteUsersSheet';
 export * from './components/InviteUsersWidget';
-export * from './components/ListItem';
+export * from './components/listItems';
 export * from './components/LongPressDisclosure';
 export * from './components/ManageChannels/ManageChannelsScreenView';
 export * from './components/MessageInput';
